backend/requirements.txt
Flask
Flask-cors
gunicorn
gevent
flask-sqlalchemy
pymysql